Crate wgpu_core_remote

This crate is a remoting version of [wgpu_core] for browser implementing WebGPU.

Modern browsers have content processes that are untrusted and cannot directly access the GPU. Instead, they communicate with a trusted GPU process that has access to the GPU.

wgpu-core-remote provides a remoting wrapper around [wgpu_core] named Global, that is used in the GPU process to execute commands from the content process. All stuff that is needed in the content process is provided in the [wgpu_core_remote_types] crate.

The key part of remoting are Id which are described in crate::hub.

Modules

global

hub

Allocating resource ids, and tracking the resources they refer to.

id

registry

storage