wgpu_core_remote::global

Struct Global

pub struct Global {
    pub(crate) hub: RefCell<Hub>,
    pub instance: Arc<Instance>,
}

Wrapper around Instance that uses Hub to store all resources created by the instance behind an Id.

All resource methods are implemented on Global and accept types from wgpu_core_remote_types (which use Ids instead of concrete resources) and maps them into wgpu_core types (the process which also includes resolving the IDs).

Fields
hub: RefCell<Hub>instance: Arc<Instance>
Implementations

impl Global

pub fn render_bundle_encoder_set_bind_group( &self, bundle_encoder: RenderBundleEncoderId, index: u32, bind_group_id: Option<BindGroupId>, offsets: &[DynamicOffset], )

pub fn render_bundle_encoder_set_pipeline( &self, bundle_encoder: RenderBundleEncoderId, pipeline_id: RenderPipelineId, )

pub fn render_bundle_encoder_set_vertex_buffer( &self, bundle_encoder: RenderBundleEncoderId, slot: u32, buffer_id: Option<BufferId>, offset: BufferAddress, size: Option<BufferSize>, )

pub fn render_bundle_encoder_set_index_buffer( &self, bundle_encoder: RenderBundleEncoderId, buffer: BufferId, index_format: IndexFormat, offset: BufferAddress, size: Option<BufferSize>, )

pub fn render_bundle_encoder_set_immediates( &self, bundle_encoder: RenderBundleEncoderId, offset: u32, data: &[u8], )

pub fn render_bundle_encoder_draw( &self, bundle_encoder: RenderBundleEncoderId, vertex_count: u32, instance_count: u32, first_vertex: u32, first_instance: u32, )

pub fn render_bundle_encoder_draw_indexed( &self, bundle_encoder: RenderBundleEncoderId, index_count: u32, instance_count: u32, first_index: u32, base_vertex: i32, first_instance: u32, )

pub fn render_bundle_encoder_draw_indirect( &self, bundle_encoder: RenderBundleEncoderId, buffer_id: BufferId, offset: BufferAddress, )

pub fn render_bundle_encoder_draw_indexed_indirect( &self, bundle_encoder: RenderBundleEncoderId, buffer_id: BufferId, offset: BufferAddress, )

pub fn render_bundle_encoder_push_debug_group( &self, bundle_encoder: RenderBundleEncoderId, label: &str, )

pub fn render_bundle_encoder_pop_debug_group( &self, bundle_encoder: RenderBundleEncoderId, )

pub fn render_bundle_encoder_insert_debug_marker( &self, bundle_encoder: RenderBundleEncoderId, label: &str, )

impl Global

pub fn handle_render_bundle_encoder_command<'a>( &self, bundle_encoder: RenderBundleEncoderId, command: RenderBundleEncoderCommand<'a>, )

impl Global

pub fn command_encoder_finish( &self, encoder_id: CommandEncoderId, desc: &CommandBufferDescriptor<'_>, id_in: CommandBufferId, )

Finishes a command encoder, creating a command buffer and returning errors that were deferred until now.

The returned String is the label of the command encoder, supplied so that wgpu can include the label when printing deferred errors without having its own copy of the label. This is a kludge and should be replaced if we think of a better solution to propagating labels.

pub fn command_encoder_push_debug_group( &self, encoder_id: CommandEncoderId, label: &str, )

pub fn command_encoder_insert_debug_marker( &self, encoder_id: CommandEncoderId, label: &str, )

pub fn command_encoder_pop_debug_group(&self, encoder_id: CommandEncoderId)

impl Global

pub fn command_encoder_clear_buffer( &self, command_encoder_id: CommandEncoderId, dst: BufferId, offset: BufferAddress, size: Option<BufferAddress>, )

pub fn command_encoder_clear_texture( &self, command_encoder_id: CommandEncoderId, dst: TextureId, subresource_range: &ImageSubresourceRange, )

impl Global

pub fn command_encoder_write_timestamp( &self, command_encoder_id: CommandEncoderId, query_set_id: QuerySetId, query_index: u32, )

pub fn command_encoder_resolve_query_set( &self, command_encoder_id: CommandEncoderId, query_set_id: QuerySetId, start_query: u32, query_count: u32, destination: BufferId, destination_offset: BufferAddress, )

impl Global

pub fn command_encoder_copy_buffer_to_buffer( &self, command_encoder_id: CommandEncoderId, source: BufferId, source_offset: BufferAddress, destination: BufferId, destination_offset: BufferAddress, size: Option<BufferAddress>, )

pub fn command_encoder_copy_buffer_to_texture( &self, command_encoder_id: CommandEncoderId, source: &TexelCopyBufferInfo, destination: &TexelCopyTextureInfo, copy_size: &Extent3d, )

pub fn command_encoder_copy_texture_to_buffer( &self, command_encoder_id: CommandEncoderId, source: &TexelCopyTextureInfo, destination: &TexelCopyBufferInfo, copy_size: &Extent3d, )

pub fn command_encoder_copy_texture_to_texture( &self, command_encoder_id: CommandEncoderId, source: &TexelCopyTextureInfo, destination: &TexelCopyTextureInfo, copy_size: &Extent3d, )

impl Global

pub fn handle_command_encoder_command<'a>( &self, command_encoder_id: CommandEncoderId, command: CommandEncoderCommand<'a>, )

impl Global

pub fn command_encoder_begin_compute_pass<'a>( &self, encoder_id: CommandEncoderId, desc: &ComputePassDescriptor<'a>, id_in: ComputePassEncoderId, )

Creates a compute pass.

If creation fails, an invalid pass is returned. Attempting to record commands into an invalid pass is permitted, but a validation error will ultimately be generated when the parent encoder is finished, and it is not possible to run any commands from the invalid pass.

If successful, puts the encoder into the Locked state.

pub fn compute_pass_end(&self, pass_id: ComputePassEncoderId)

pub fn compute_pass_remove(&self, pass_id: ComputePassEncoderId) -> ComputePass

impl Global

pub fn compute_pass_set_bind_group( &self, pass_id: ComputePassEncoderId, index: u32, bind_group_id: Option<BindGroupId>, offsets: &[DynamicOffset], )

pub fn compute_pass_set_pipeline( &self, pass_id: ComputePassEncoderId, pipeline_id: ComputePipelineId, )

pub fn compute_pass_set_immediates( &self, pass_id: ComputePassEncoderId, offset: u32, data: &[u8], )

pub fn compute_pass_dispatch_workgroups( &self, pass_id: ComputePassEncoderId, groups_x: u32, groups_y: u32, groups_z: u32, )

pub fn compute_pass_dispatch_workgroups_indirect( &self, pass_id: ComputePassEncoderId, buffer_id: BufferId, offset: BufferAddress, )

pub fn compute_pass_push_debug_group( &self, pass_id: ComputePassEncoderId, label: &str, color: u32, )

pub fn compute_pass_pop_debug_group(&self, pass_id: ComputePassEncoderId)

pub fn compute_pass_insert_debug_marker( &self, pass_id: ComputePassEncoderId, label: &str, color: u32, )

pub fn compute_pass_write_timestamp( &self, pass_id: ComputePassEncoderId, query_set_id: QuerySetId, query_index: u32, )

pub fn compute_pass_begin_pipeline_statistics_query( &self, pass_id: ComputePassEncoderId, query_set_id: QuerySetId, query_index: u32, )

pub fn compute_pass_end_pipeline_statistics_query( &self, pass_id: ComputePassEncoderId, )

impl Global

pub fn handle_compute_pass_command( &self, pass_id: ComputePassEncoderId, command: ComputePassEncoderCommand, )

impl Global

pub fn device_features(&self, device_id: DeviceId) -> Features

pub fn device_limits(&self, device_id: DeviceId) -> Limits

pub fn device_adapter_info(&self, device_id: DeviceId) -> AdapterInfo

pub fn device_downlevel_properties( &self, device_id: DeviceId, ) -> DownlevelCapabilities

pub fn device_create_buffer( &self, device_id: DeviceId, desc: &BufferDescriptor<'_>, id_in: BufferId, )

pub fn create_buffer_error( &self, device_id: DeviceId, id_in: BufferId, desc: &BufferDescriptor<'_>, )

Assign id_in an error with the given label.

Ensure that future attempts to use id_in as a buffer ID will propagate the error, following the WebGPU “contagious invalidity” style.

Firefox uses this function to comply strictly with the WebGPU spec, which requires GPUBufferDescriptor validation to be generated on the Device timeline and leave the newly created GPUBuffer invalid.

Ideally, we would simply let Device::create_buffer take care of all of this, but some errors must be detected before we can even construct a wgpu_types::BufferDescriptor to give it. For example, the WebGPU API allows a GPUBufferDescriptor’s usage property to be any WebIDL unsigned long value, but we can’t construct a wgpu_types::BufferUsages value from values with unassigned bits set. This means we must validate usage before we can call Device::create_buffer.

When that validation fails, we must arrange for the buffer id to be considered invalid. This method provides the means to do so.

pub fn create_render_bundle_error( &self, device_id: DeviceId, id_in: RenderBundleId, desc: &RenderBundleDescriptor<'_>, )

Assign id_in an error with the given label.

See Self::create_buffer_error for more context and explanation.

pub fn create_texture_error( &self, device_id: DeviceId, id_in: TextureId, desc: &TextureDescriptor<'_>, ) -> TextureId

Assign id_in an error with the given label.

See Self::create_buffer_error for more context and explanation.

pub fn create_external_texture_error( &self, device_id: DeviceId, id_in: ExternalTextureId, desc: &ExternalTextureDescriptor<'_>, )

Assign id_in an error with the given label.

See Self::create_buffer_error for more context and explanation.

pub fn create_bind_group_layout_error( &self, device_id: DeviceId, id_in: BindGroupLayoutId, label: Option<Cow<'_, str>>, )

Assign id_in an error with the given label.

In JavaScript environments, it is possible to call GPUDevice.createBindGroupLayout with entries that are invalid. Because our Rust’s types for bind group layouts prevent even calling Self::device_create_bind_group, we let standards-compliant environments register an invalid bind group layout so this crate’s API can still be consistently used.

See Self::create_buffer_error for additional context and explanation.

pub fn buffer_destroy(&self, buffer_id: BufferId)

pub fn buffer_remove(&self, buffer_id: BufferId) -> Arc<Buffer>

pub fn device_create_texture( &self, device_id: DeviceId, desc: &TextureDescriptor<'_>, id_in: TextureId, )

pub fn device_validate_texture_descriptor( &self, device_id: DeviceId, desc: &TextureDescriptor<'_>, ) -> Option<CreateTextureError>

pub unsafe fn create_texture_from_hal( &self, hal_texture: Box<dyn DynTexture>, device_id: DeviceId, desc: &TextureDescriptor<'_>, initial_state: TextureUses, id_in: TextureId, cleared: bool, ) -> (TextureId, Option<CreateTextureError>)

Safety

hal_texture must be created from device_id corresponding raw handle.

hal_texture must be created respecting desc

hal_texture must be initialized

The initial_state must match the actual driver-side state of the wrapped resource at the moment of wrap.

pub unsafe fn create_buffer_from_hal<A: Api>( &self, hal_buffer: A::Buffer, device_id: DeviceId, desc: &BufferDescriptor<'_>, id_in: BufferId, ) -> (BufferId, Option<CreateBufferError>)

Safety

hal_buffer must be created from device_id corresponding raw handle.

hal_buffer must be created respecting desc

hal_buffer must be initialized

hal_buffer must not have zero size.

pub fn texture_destroy(&self, texture_id: TextureId)

pub fn texture_remove(&self, texture_id: TextureId) -> Arc<Texture>

pub unsafe fn texture_mark_externally_initialized(&self, texture_id: TextureId)

Safety

The entire contents of the texture must already be initialized.

pub fn texture_create_view( &self, texture_id: TextureId, desc: &TextureViewDescriptor<'_>, id_in: TextureViewId, )

pub fn texture_view_remove( &self, texture_view_id: TextureViewId, ) -> Arc<TextureView>

pub fn device_create_external_texture( &self, device_id: DeviceId, desc: &ExternalTextureDescriptor<'_>, planes: &[TextureViewId], id_in: ExternalTextureId, )

pub fn external_texture_destroy(&self, external_texture_id: ExternalTextureId)

pub fn external_texture_remove( &self, external_texture_id: ExternalTextureId, ) -> Arc<ExternalTexture>

pub fn device_create_sampler( &self, device_id: DeviceId, desc: &SamplerDescriptor<'_>, id_in: SamplerId, )

pub fn sampler_remove(&self, sampler_id: SamplerId) -> Arc<Sampler>

pub fn device_create_bind_group_layout( &self, device_id: DeviceId, desc: &BindGroupLayoutDescriptor<'_>, id_in: BindGroupLayoutId, )

pub fn bind_group_layout_remove( &self, bind_group_layout_id: BindGroupLayoutId, ) -> Arc<BindGroupLayout>

pub fn device_create_pipeline_layout( &self, device_id: DeviceId, desc: &PipelineLayoutDescriptor<'_>, id_in: PipelineLayoutId, )

pub fn pipeline_layout_remove( &self, pipeline_layout_id: PipelineLayoutId, ) -> Arc<PipelineLayout>

pub fn device_create_bind_group( &self, device_id: DeviceId, desc: &BindGroupDescriptor<'_>, id_in: BindGroupId, )

pub fn bind_group_remove(&self, bind_group_id: BindGroupId) -> Arc<BindGroup>

pub fn device_create_shader_module( &self, device_id: DeviceId, desc: &ShaderModuleDescriptor<'_>, id_in: ShaderModuleId, )

Create a shader module with the given source.

This function may consume a lot of stack space. Compiler-enforced limits for parsing recursion exist; if shader compilation runs into them, it will return an error gracefully. However, on some build profiles and platforms, the default stack size for a thread may be exceeded before this limit is reached during parsing. Callers should ensure that there is enough stack space for this, particularly if calls to this method are exposed to user input.

pub fn shader_module_compilation_info( &self, shader_module_id: ShaderModuleId, ) -> CompilationInfo

pub fn shader_module_remove( &self, shader_module_id: ShaderModuleId, ) -> Arc<ShaderModule>

pub fn device_create_command_encoder( &self, device_id: DeviceId, desc: &CommandEncoderDescriptor<Label<'_>>, id_in: CommandEncoderId, )

pub fn command_encoder_remove( &self, command_encoder_id: CommandEncoderId, ) -> Arc<CommandEncoder>

pub fn command_buffer_remove( &self, command_buffer_id: CommandBufferId, ) -> Arc<CommandBuffer>

pub fn device_create_render_bundle_encoder( &self, device_id: DeviceId, desc: &RenderBundleEncoderDescriptor<'_>, id_in: RenderBundleEncoderId, )

pub fn render_bundle_encoder_finish( &self, render_bundle_encoder_id: RenderBundleEncoderId, desc: RenderBundleDescriptor<'_>, id_in: RenderBundleId, )

pub fn render_bundle_encoder_remove( &self, render_bundle_encoder_id: RenderBundleEncoderId, ) -> RenderBundleEncoder

pub fn render_bundle_remove( &self, render_bundle_id: RenderBundleId, ) -> Arc<RenderBundle>

pub fn device_create_query_set( &self, device_id: DeviceId, desc: &QuerySetDescriptor<'_>, id_in: QuerySetId, )

pub fn query_set_destroy(&self, query_set_id: QuerySetId)

pub fn query_set_remove(&self, query_set_id: QuerySetId) -> Arc<QuerySet>

fn resolve_render_pipeline_descriptor<'a>( shader_modules: &mut Registry<Arc<ShaderModule>>, pipeline_layouts: &mut Registry<Arc<PipelineLayout>>, desc: &'a RenderPipelineDescriptor<'_>, ) -> ResolvedGeneralRenderPipelineDescriptor<'a>

pub fn device_create_render_pipeline( &self, device_id: DeviceId, desc: &RenderPipelineDescriptor<'_>, id_in: RenderPipelineId, )

pub fn create_render_pipeline_or_error( &self, device_id: DeviceId, desc: &RenderPipelineDescriptor<'_>, id_in: RenderPipelineId, ) -> Result<(), CreateRenderPipelineError>

Error-returning version of device_create_render_pipeline to implement GPUDevice.createRenderPipelineAsync. Returns an error if the pipeline creation fails instead of handling error in device.

Id is assigned to the pipeline only if the creation succeeds.

pub fn render_pipeline_get_bind_group_layout( &self, pipeline_id: RenderPipelineId, index: u32, id_in: BindGroupLayoutId, )

Get an ID of one of the bind group layouts. The ID adds a refcount, which needs to be released by calling bind_group_layout_drop.

pub fn render_pipeline_remove( &self, render_pipeline_id: RenderPipelineId, ) -> Arc<RenderPipeline>

pub fn device_create_compute_pipeline( &self, device_id: DeviceId, desc: &ComputePipelineDescriptor<'_>, id_in: ComputePipelineId, )

pub fn device_create_compute_pipeline_or_error( &self, device_id: DeviceId, desc: &ComputePipelineDescriptor<'_>, id_in: ComputePipelineId, ) -> Result<(), CreateComputePipelineError>

Error-returning version of device_create_compute_pipeline to implement GPUDevice.createComputePipelineAsync. Returns an error if the pipeline creation fails instead of handling error in device.

Id is assigned to the pipeline only if the creation succeeds.

pub fn compute_pipeline_get_bind_group_layout( &self, pipeline_id: ComputePipelineId, index: u32, id_in: BindGroupLayoutId, )

Get an ID of one of the bind group layouts. The ID adds a refcount, which needs to be released by calling bind_group_layout_drop.

pub fn compute_pipeline_remove( &self, compute_pipeline_id: ComputePipelineId, ) -> Arc<ComputePipeline>

pub fn device_poll( &self, device_id: DeviceId, poll_type: PollType<SubmissionIndex>, ) -> Result<PollStatus, WaitIdleError>

Check device_id for freeable resources and completed buffer mappings.

pub fn poll_all_devices(&self, force_wait: bool) -> Result<bool, WaitIdleError>

Poll all devices on all backends.

This is the implementation of wgpu::Instance::poll_all.

Return all_queue_empty indicating whether there are more queue submissions still in flight.

pub unsafe fn device_start_graphics_debugger_capture(&self, device_id: DeviceId)

Safety

See wgpu::Device::start_graphics_debugger_capture for details the safety.

pub unsafe fn device_stop_graphics_debugger_capture(&self, device_id: DeviceId)

Safety

See wgpu::Device::stop_graphics_debugger_capture for details the safety.

pub fn device_remove(&self, device_id: DeviceId) -> Arc<Device>

pub fn device_set_device_lost_closure( &self, device_id: DeviceId, device_lost_closure: DeviceLostClosure, )

device_lost_closure might never be called.

pub fn device_destroy(&self, device_id: DeviceId)

pub fn device_get_internal_counters( &self, device_id: DeviceId, ) -> InternalCounters

pub fn device_generate_allocator_report( &self, device_id: DeviceId, ) -> Option<AllocatorReport>

pub fn queue_remove(&self, queue_id: QueueId) -> Arc<Queue>

pub fn buffer_map_async( &self, buffer_id: BufferId, offset: BufferAddress, size: Option<BufferAddress>, op: BufferMapOperation, ) -> Option<SubmissionIndex>

op.callback is always called, even in case of errors.

pub fn buffer_get_mapped_range( &self, buffer_id: BufferId, offset: BufferAddress, size: Option<BufferAddress>, ) -> Result<(NonNull<u8>, u64), BufferAccessError>

pub fn buffer_unmap(&self, buffer_id: BufferId)

pub fn device_on_uncaptured_error( &self, device_id: DeviceId, handler: Arc<dyn UncapturedErrorHandler>, )

pub fn device_push_error_scope( &self, device_id: DeviceId, error_scope: ErrorFilter, )

pub fn device_pop_error_scope( &self, device_id: DeviceId, ) -> Result<Option<Error>, EmptyErrorScopeStack>

pub fn device_handle_error( &self, device_id: DeviceId, source: impl WebGpuError + Send + Sync + 'static, label: Option<&str>, fn_ident: &'static str, )

impl Global

pub fn request_adapter( &self, desc: &RequestAdapterOptions, apply_limit_buckets: bool, backends: Backends, id_in: AdapterId, ) -> Result<AdapterId, RequestAdapterError>

pub unsafe fn create_adapter_from_hal( &self, hal_adapter: DynExposedAdapter, id_in: AdapterId, ) -> AdapterId

Create an adapter from a HAL adapter.

The HAL adapter may be obtained e.g. by calling enumerate_adapters on the HAL directly.

If limit bucketing is desired, wgpu_core::limits::apply_limit_buckets should be called with the HAL adapter before calling this function.

Safety

hal_adapter must be created from this global internal instance handle.

pub fn adapter_get_info(&self, adapter_id: AdapterId) -> AdapterInfo

pub fn adapter_get_texture_format_features( &self, adapter_id: AdapterId, format: TextureFormat, ) -> TextureFormatFeatures

pub fn adapter_features(&self, adapter_id: AdapterId) -> Features

pub fn adapter_limits(&self, adapter_id: AdapterId) -> Limits

pub fn adapter_downlevel_capabilities( &self, adapter_id: AdapterId, ) -> DownlevelCapabilities

pub fn adapter_get_presentation_timestamp( &self, adapter_id: AdapterId, ) -> PresentationTimestamp

pub fn adapter_cooperative_matrix_properties( &self, adapter_id: AdapterId, ) -> Vec<CooperativeMatrixProperties>

pub fn adapter_remove(&self, adapter_id: AdapterId) -> Arc<Adapter>

impl Global

pub fn adapter_request_device( &self, adapter_id: AdapterId, desc: &DeviceDescriptor<'_>, device_id_in: DeviceId, queue_id_in: QueueId, ) -> Result<(DeviceId, QueueId), RequestDeviceError>

pub fn adapter_validate_device_descriptor( &self, adapter_id: AdapterId, desc: &mut DeviceDescriptor<'_>, ) -> Result<(), RequestDeviceError>

pub unsafe fn create_device_from_hal( &self, adapter_id: AdapterId, hal_device: DynOpenDevice, desc: &DeviceDescriptor<'_>, device_id_in: DeviceId, queue_id_in: QueueId, ) -> Result<(DeviceId, QueueId), RequestDeviceError>

Safety

hal_device must be created from adapter_id or its internal handle.

desc must be a subset of hal_device features and limits.

impl Global

pub fn queue_write_buffer( &self, queue_id: QueueId, buffer_id: BufferId, buffer_offset: BufferAddress, data: &[u8], )

pub fn queue_write_texture( &self, queue_id: QueueId, destination: &TexelCopyTextureInfo, data: &[u8], data_layout: &TexelCopyBufferLayout, size: &Extent3d, )

pub fn queue_submit( &self, queue_id: QueueId, command_buffer_ids: &[CommandBufferId], ) -> SubmissionIndex

pub fn queue_on_submitted_work_done( &self, queue_id: QueueId, closure: SubmittedWorkDoneClosure, ) -> SubmissionIndex

impl Global

pub fn command_encoder_begin_render_pass<'a>( &self, encoder_id: CommandEncoderId, desc: &RenderPassDescriptor<'a>, id_in: RenderPassEncoderId, )

Creates a render pass.

If creation fails, an invalid pass is returned. Attempting to record commands into an invalid pass is permitted, but a validation error will ultimately be generated when the parent encoder is finished, and it is not possible to run any commands from the invalid pass.

If successful, puts the encoder into the Locked state.

pub fn render_pass_end(&self, pass: RenderPassEncoderId)

pub fn render_pass_remove(&self, pass: RenderPassEncoderId) -> RenderPass

impl Global

pub fn render_pass_set_bind_group( &self, pass: RenderPassEncoderId, index: u32, bind_group_id: Option<BindGroupId>, offsets: &[DynamicOffset], )

pub fn render_pass_set_pipeline( &self, pass: RenderPassEncoderId, pipeline_id: RenderPipelineId, )

pub fn render_pass_set_index_buffer( &self, pass: RenderPassEncoderId, buffer_id: BufferId, index_format: IndexFormat, offset: BufferAddress, size: Option<BufferSize>, )

pub fn render_pass_set_vertex_buffer( &self, pass: RenderPassEncoderId, slot: u32, buffer_id: Option<BufferId>, offset: BufferAddress, size: Option<BufferSize>, )

pub fn render_pass_set_blend_constant( &self, pass: RenderPassEncoderId, color: Color, )

pub fn render_pass_set_stencil_reference( &self, pass: RenderPassEncoderId, value: u32, )

pub fn render_pass_set_viewport( &self, pass: RenderPassEncoderId, x: f32, y: f32, w: f32, h: f32, depth_min: f32, depth_max: f32, )

pub fn render_pass_set_scissor_rect( &self, pass: RenderPassEncoderId, x: u32, y: u32, w: u32, h: u32, )

pub fn render_pass_set_immediates( &self, pass: RenderPassEncoderId, offset: u32, data: &[u8], )

pub fn render_pass_draw( &self, pass: RenderPassEncoderId, vertex_count: u32, instance_count: u32, first_vertex: u32, first_instance: u32, )

pub fn render_pass_draw_indexed( &self, pass: RenderPassEncoderId, index_count: u32, instance_count: u32, first_index: u32, base_vertex: i32, first_instance: u32, )

pub fn render_pass_draw_indirect( &self, pass: RenderPassEncoderId, buffer_id: BufferId, offset: BufferAddress, )

pub fn render_pass_draw_indexed_indirect( &self, pass: RenderPassEncoderId, buffer_id: BufferId, offset: BufferAddress, )

pub fn render_pass_push_debug_group( &self, pass: RenderPassEncoderId, label: &str, color: u32, )

pub fn render_pass_pop_debug_group(&self, pass: RenderPassEncoderId)

pub fn render_pass_insert_debug_marker( &self, pass: RenderPassEncoderId, label: &str, color: u32, )

pub fn render_pass_write_timestamp( &self, pass: RenderPassEncoderId, query_set_id: QuerySetId, query_index: u32, )

pub fn render_pass_begin_occlusion_query( &self, pass: RenderPassEncoderId, query_index: u32, )

pub fn render_pass_end_occlusion_query(&self, pass: RenderPassEncoderId)

pub fn render_pass_begin_pipeline_statistics_query( &self, pass: RenderPassEncoderId, query_set_id: QuerySetId, query_index: u32, )

pub fn render_pass_end_pipeline_statistics_query( &self, pass: RenderPassEncoderId, )

pub fn render_pass_execute_bundles( &self, pass: RenderPassEncoderId, render_bundle_ids: &[RenderBundleId], )

impl Global

pub fn handle_render_pass_command( &self, pass_id: RenderPassEncoderId, command: RenderPassEncoderCommand, )

impl Global

pub fn new( name: &str, instance_desc: InstanceDescriptor, telemetry: Option<Telemetry>, ) -> Self

pub fn from_instance(instance: Arc<Instance>) -> Self

pub fn instance(&self) -> &Arc<Instance>

impl Global

pub fn import_adapter( &self, adapter: Arc<Adapter>, id_in: AdapterId, ) -> AdapterId

Import Arc<Adapter> into the global hub, returning an AdapterId under which the adapter is stored.

pub fn resolve_adapter_id(&self, adapter_id: AdapterId) -> Arc<Adapter>

Resolve an AdapterId to the corresponding Arc<Adapter> in the global hub.

pub fn import_device(&self, device: Arc<Device>, id_in: DeviceId) -> DeviceId

Import Arc<Device> into the global hub, returning a DeviceId under which the device is stored.

pub fn resolve_device_id(&self, device_id: DeviceId) -> Arc<Device>

Resolve a DeviceId to the corresponding Arc<Device> in the global hub.

pub fn import_queue(&self, queue: Arc<Queue>, id_in: QueueId) -> QueueId

Import Arc<Queue> into the global hub, returning a QueueId under which the queue is stored.

pub fn resolve_queue_id(&self, queue_id: QueueId) -> Arc<Queue>

Resolve a QueueId to the corresponding Arc<Queue> in the global hub.

pub fn import_pipeline_layout( &self, pipeline_layout: Arc<PipelineLayout>, id_in: PipelineLayoutId, ) -> PipelineLayoutId

Import Arc<PipelineLayout> into the global hub, returning a PipelineLayoutId under which the pipeline layout is stored.

pub fn resolve_pipeline_layout_id( &self, pipeline_layout_id: PipelineLayoutId, ) -> Arc<PipelineLayout>

Resolve a PipelineLayoutId to the corresponding Arc<PipelineLayout> in the global hub.

pub fn import_shader_module( &self, shader_module: Arc<ShaderModule>, id_in: ShaderModuleId, ) -> ShaderModuleId

Import Arc<ShaderModule> into the global hub, returning a ShaderModuleId under which the shader module is stored.

pub fn resolve_shader_module_id( &self, shader_module_id: ShaderModuleId, ) -> Arc<ShaderModule>

Resolve a ShaderModuleId to the corresponding Arc<ShaderModule> in the global hub.

pub fn import_bind_group_layout( &self, bind_group_layout: Arc<BindGroupLayout>, id_in: BindGroupLayoutId, ) -> BindGroupLayoutId

Import Arc<BindGroupLayout> into the global hub, returning a BindGroupLayoutId under which the bind group layout is stored.

pub fn resolve_bind_group_layout_id( &self, bind_group_layout_id: BindGroupLayoutId, ) -> Arc<BindGroupLayout>

Resolve a BindGroupLayoutId to the corresponding Arc<BindGroupLayout> in the global hub.

pub fn import_bind_group( &self, bind_group: Arc<BindGroup>, id_in: BindGroupId, ) -> BindGroupId

Import Arc<BindGroup> into the global hub, returning a BindGroupId under which the bind group is stored.

pub fn resolve_bind_group_id( &self, bind_group_id: BindGroupId, ) -> Arc<BindGroup>

Resolve a BindGroupId to the corresponding Arc<BindGroup> in the global hub.

pub fn import_command_encoder( &self, command_encoder: Arc<CommandEncoder>, id_in: CommandEncoderId, ) -> CommandEncoderId

Import Arc<CommandEncoder> into the global hub, returning a CommandEncoderId under which the command encoder is stored.

pub fn resolve_command_encoder_id( &self, command_encoder_id: CommandEncoderId, ) -> Arc<CommandEncoder>

Resolve a CommandEncoderId to the corresponding Arc<CommandEncoder> in the global hub.

pub fn import_command_buffer( &self, command_buffer: Arc<CommandBuffer>, id_in: CommandBufferId, ) -> CommandBufferId

Import Arc<CommandBuffer> into the global hub, returning a CommandBufferId under which the command buffer is stored.

pub fn resolve_command_buffer_id( &self, command_buffer_id: CommandBufferId, ) -> Arc<CommandBuffer>

Resolve a CommandBufferId to the corresponding Arc<CommandBuffer> in the global hub.

pub fn import_render_bundle( &self, render_bundle: Arc<RenderBundle>, id_in: RenderBundleId, ) -> RenderBundleId

Import Arc<RenderBundle> into the global hub, returning a RenderBundleId under which the render bundle is stored.

pub fn resolve_render_bundle_id( &self, render_bundle_id: RenderBundleId, ) -> Arc<RenderBundle>

Resolve a RenderBundleId to the corresponding Arc<RenderBundle> in the global hub.

pub fn import_render_pipeline( &self, render_pipeline: Arc<RenderPipeline>, id_in: RenderPipelineId, ) -> RenderPipelineId

Import Arc<RenderPipeline> into the global hub, returning a RenderPipelineId under which the render pipeline is stored.

pub fn resolve_render_pipeline_id( &self, render_pipeline_id: RenderPipelineId, ) -> Arc<RenderPipeline>

Resolve a RenderPipelineId to the corresponding Arc<RenderPipeline> in the global hub.

pub fn import_compute_pipeline( &self, compute_pipeline: Arc<ComputePipeline>, id_in: ComputePipelineId, ) -> ComputePipelineId

Import Arc<ComputePipeline> into the global hub, returning a ComputePipelineId under which the compute pipeline is stored.

pub fn resolve_compute_pipeline_id( &self, compute_pipeline_id: ComputePipelineId, ) -> Arc<ComputePipeline>

Resolve a ComputePipelineId to the corresponding Arc<ComputePipeline> in the global hub.

pub fn import_query_set( &self, query_set: Arc<QuerySet>, id_in: QuerySetId, ) -> QuerySetId

Import Arc<QuerySet> into the global hub, returning a QuerySetId under which the query set is stored.

pub fn resolve_query_set_id(&self, query_set_id: QuerySetId) -> Arc<QuerySet>

Resolve a QuerySetId to the corresponding Arc<QuerySet> in the global hub.

pub fn import_buffer(&self, buffer: Arc<Buffer>, id_in: BufferId) -> BufferId

Import Arc<Buffer> into the global hub, returning a BufferId under which the buffer is stored.

pub fn resolve_buffer_id(&self, buffer_id: BufferId) -> Arc<Buffer>

Resolve a BufferId to the corresponding Arc<Buffer> in the global hub.

pub fn import_texture( &self, texture: Arc<Texture>, id_in: TextureId, ) -> TextureId

Import Arc<Texture> into the global hub, returning a TextureId under which the texture is stored.

pub fn resolve_texture_id(&self, texture_id: TextureId) -> Arc<Texture>

Resolve a TextureId to the corresponding Arc<Texture> in the global hub.

pub fn import_texture_view( &self, texture_view: Arc<TextureView>, id_in: TextureViewId, ) -> TextureViewId

Import Arc<TextureView> into the global hub, returning a TextureViewId under which the texture view is stored.

pub fn resolve_texture_view_id( &self, texture_view_id: TextureViewId, ) -> Arc<TextureView>

Resolve a TextureViewId to the corresponding Arc<TextureView> in the global hub.

pub fn import_external_texture( &self, external_texture: Arc<ExternalTexture>, id_in: ExternalTextureId, ) -> ExternalTextureId

Import Arc<ExternalTexture> into the global hub, returning a ExternalTextureId under which the external texture is stored.

pub fn resolve_external_texture_id( &self, external_texture_id: ExternalTextureId, ) -> Arc<ExternalTexture>

Resolve a ExternalTextureId to the corresponding Arc<ExternalTexture> in the global hub.

pub fn import_sampler( &self, sampler: Arc<Sampler>, id_in: SamplerId, ) -> SamplerId

Import Arc<Sampler> into the global hub, returning a SamplerId under which the sampler is stored.

pub fn resolve_sampler_id(&self, sampler_id: SamplerId) -> Arc<Sampler>

Resolve a SamplerId to the corresponding Arc<Sampler> in the global hub.

pub fn import_render_pass( &self, render_pass: RenderPass, id_in: RenderPassEncoderId, ) -> RenderPassEncoderId

Import RenderPass into the global hub, returning a RenderPassEncoderId under which the render pass is stored.

pub fn import_compute_pass( &self, compute_pass: ComputePass, id_in: ComputePassEncoderId, ) -> ComputePassEncoderId

Import ComputePass into the global hub, returning a ComputePassEncoderId under which the compute pass is stored.

pub fn import_render_bundle_encoder( &self, render_bundle_encoder: RenderBundleEncoder, id_in: RenderBundleEncoderId, ) -> RenderBundleEncoderId

Import RenderBundleEncoder into the global hub, returning a RenderBundleEncoderId under which the render bundle encoder is stored.