wgpu_core_remote

Module hub

Allocating resource ids, and tracking the resources they refer to.

The wgpu_core API uses identifiers of type Id<R> to refer to resources of type R. For example, id::DeviceId is an alias for Id<markers::Device>, and id::BufferId is an alias for Id<markers::Buffer>. Id implements Copy, Hash, Eq, Ord, and of course Debug.

Ids also incorporate a generation number, for additional validation.

The resources to which identifiers refer are freed explicitly. Attempting to use an identifier for a resource that has been freed elicits an panic.

Assigning ids to resources

Firefox and Servo allocate ids themselves in the content process using IdentityHub, then pass then via IPC to GPU process where it’s passed down to Global::device_create_buffer and friends the id to assign the new resource.

Methods that create resources all expect an id_in argument that the caller uses to specify the id. For example, the declaration of Global::device_create_buffer looks like this:

impl Global {
    /* ... */
    pub fn device_create_buffer(
        &self,
        device_id: id::DeviceId,
        desc: &BufferDescriptor,
        id_in: id::BufferId,
    ) {
        /* ... */
    }
    /* ... */
}

Id allocation and streaming

Allowing users to assign resource ids themselves enables major performance improvements in some applications.

This is designed for use by Firefox’s and Servo’s WebGPU implementation. For security, web content and GPU use must be kept segregated in separate processes, with all interaction between them mediated by an inter-process communication protocol. As web content uses the WebGPU API, the content process sends messages to the GPU process, which interacts with the platform’s GPU APIs on content’s behalf, occasionally sending results back.

In a classic Rust API, a resource allocation function takes parameters describing the resource to create, and if creation succeeds, it returns the resource id in a Result::Ok value. However, this design is a poor fit for the split-process design described above: content must wait for the reply to its buffer-creation message (say) before it can know which id it can use in the next message that uses that buffer. On a common usage pattern, the classic Rust design imposes the latency of a full cross-process round trip.

We can avoid incurring these round-trip latencies simply by letting the content process assign resource ids itself. With this approach, content can choose an id for the new buffer, send a message to create the buffer, and then immediately send the next message operating on that buffer, since it already knows its id. Allowing content and GPU process activity to be pipelined greatly improves throughput.

To help propagate errors correctly in this style of usage, when resource creation fails, the id supplied for that resource is marked to indicate as much, allowing subsequent operations using that id to be properly flagged as errors as well. This is called Contagious Invalidity.

Structs

Hub

All the resources tracked by a crate::global::Global.